+++ Cargo.nix
@@ -7711,7 +7711,7 @@ rec {
           {
             name = "nix";
             packageId = "nix 0.31.3";
-            features = [ "fs" "process" "signal" ];
+            features = [ "fs" "process" "signal" "term" ];
           }
           {
             name = "once_cell";
@@ -8043,7 +8043,7 @@ rec {
             name = "nix";
             packageId = "nix 0.31.3";
             target = { target, features }: (target."unix" or false);
-            features = [ "fs" "process" "signal" "user" ];
+            features = [ "fs" "process" "signal" "term" "user" ];
           }
           {
             name = "nix-conf-parser";
@@ -8328,7 +8328,7 @@ rec {
           {
             name = "nix";
             packageId = "nix 0.31.3";
-            features = [ "fs" "process" "signal" ];
+            features = [ "fs" "process" "signal" "term" ];
           }
           {
             name = "nix-bindings-bindgen-raw";
@@ -8484,7 +8484,7 @@ rec {
           {
             name = "nix";
             packageId = "nix 0.31.3";
-            features = [ "fs" "process" "signal" ];
+            features = [ "fs" "process" "signal" "term" ];
           }
           {
             name = "portable-pty";
@@ -8876,7 +8876,7 @@ rec {
             name = "nix";
             packageId = "nix 0.31.3";
             target = { target, features }: (target."unix" or false);
-            features = [ "fs" "process" "signal" "user" "signal" ];
+            features = [ "fs" "process" "signal" "term" "user" "signal" ];
           }
           {
             name = "petgraph";
@@ -19021,7 +19021,7 @@ rec {
           "user" = [ "feature" ];
           "zerocopy" = [ "fs" "uio" ];
         };
-        resolvedDefaultFeatures = [ "default" "feature" "fs" "poll" "process" "signal" "user" ];
+        resolvedDefaultFeatures = [ "default" "feature" "fs" "poll" "process" "signal" "term" "user" ];
       };
       "nix-bindings-bindgen-raw" = rec {
         crateName = "nix-bindings-bindgen-raw";
@@ -32388,7 +32388,7 @@ rec {
           {
             name = "nix";
             packageId = "nix 0.31.3";
-            features = [ "fs" "process" "signal" "signal" "process" ];
+            features = [ "fs" "process" "signal" "term" "signal" "process" ];
           }
           {
             name = "tokio";